Use size_t and ssize_t for indexing strings and ports
Avoid longs, which are 4 bytes on x86_64-w64-mingw32 and cannot
index the full possible range of strings.

* libguile/strings.c (scm_string): use ssize_t for index
  (scm_string_ref): use size_t for index
  (scm_string_set_x): use size_t for index
* libguile/ports.c (update_port_position): use ssize_t for line number

+++ libguile/strings.c
@@ -1074,7 +1074,7 @@ SCM_DEFINE (scm_string, "string", 0, 0, 1,
   SCM rest;
   size_t len;
   size_t p = 0;
-  long i;
+  ssize_t i;
   int wide = 0;
 
   /* Verify that this is a list of chars.  */
@@ -1219,7 +1219,7 @@ SCM_DEFINE (scm_string_ref, "string-ref", 2, 0, 0,
 #define FUNC_NAME s_scm_string_ref
 {
   size_t len;
-  unsigned long idx;
+  size_t idx;
 
   SCM_VALIDATE_STRING (1, str);
 
@@ -1256,7 +1256,7 @@ SCM_DEFINE (scm_string_set_x, "string-set!", 3, 0, 0,
 #define FUNC_NAME s_scm_string_set_x
 {
   size_t len;
-  unsigned long idx;
+  size_t idx;
 
   SCM_VALIDATE_STRING (1, str);
